Protocol change in friendRequest
The friend request connection is now encrypted with AES using a key
determined with the DHM key exchange scheme.

The friend request connection handshake now looks like a normal
connection, only the connection type in the preable still indicates it
is a friend req connection. The connection setup goes through the same
DHM key exchange steps as in the case of a normal connection. Signatures
are exchanged, but not verified in case of friend req connection.

Finally, the client making the friend request (friendrequestee) sends
its cert, and when the core being "friend requested" accepts the
requests, it sends its cert to the friendrequestee.