Re-enabled friend requests.
The friend request now works using a normal which connection, which is
flagged as insecure by means on ctx->friend_request_connection==true.
The friend request is sent as core-to-core RPC message, and if the
request is accepted, the recepient's own certificate is sent back to the
requestor. In case of fried req declined, an RPC message with
appropriate message is sent.